[refacto] move game constants to a game module
