+++ src/game/buildings.constants.ts
