+++ src/game/game.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { GameService } from './game.service';
+
+@Module({
+  providers: [GameService],
+  exports: [GameService],
+})
+export class GameModule {}
