+++ src/game/troops.constants.ts
